fix(wu-24): address PR review — conftest import, env teardown, retry assertion
- Move run_dbt() from conftest.py to dbt_utils.py to avoid double-import
  anti-pattern (pytest auto-discovers conftest; explicit import loads twice)
- Capture AWS env var prior state before setdefault() and restore in
  teardown to prevent env-state leakage between test sessions
- Remove flawed retry assertion: dbt run without --state re-runs all
  models, so asserting selective execution was based on incorrect premise


<noreply@anthropic.com>
